update course module, full crud, link with course author user

+++ main/src/course/course.service.ts
@@ -1,7 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Course } from './course.entity';
 import { Repository } from 'typeorm';
+import { CreateCourseInput } from './input.types/create.course.input';
+import { UpdateCourseType } from './input.types/update.course.input';
 
 @Injectable()
 export class CourseService {
@@ -11,14 +13,40 @@ export class CourseService {
     ) {}
 
     async findAll(): Promise<Course[]> {
-        return this.courseRepository.find();
+        return this.courseRepository.find({ relations: ['author'] });
     }
 
     async findOne(id: number): Promise<Course> {
-        return this.courseRepository.findOne(id);
+        return this.courseRepository.findOne(id, { relations: ['author'] });
     }
 
-    async deleteOne(id: number): Promise<void> {
+    async create(createCourseData: CreateCourseInput): Promise<Course> {
+        const course = new Course();
+        course.name = createCourseData.name;
+        course.description = createCourseData.description;
+        course.headline = createCourseData.headline;
+        course.isPaid = createCourseData.isPaid;
+        course.price = createCourseData.price;
+        course.previewUrl = createCourseData.previewUrl;
+        return course;
+    }
+
+    async save(course: Course): Promise<Course> {
+        return this.courseRepository.save(course);
+    }
+
+    async updateOne(
+        id: number,
+        updateCourseData: UpdateCourseType,
+    ): Promise<Course> {
+        await this.save({ id, ...updateCourseData });
+        return this.findOne(id);
+    }
+
+    async deleteOne(id: number): Promise<Course> {
+        let course = await this.findOne(id);
+        if (!course) throw new NotFoundException('This course does not exists');
         await this.courseRepository.delete(id);
+        return course;
     }
 }
